The program now takes care of creating host directories.
The destination directory should still exist though.
Example:
    Given the destination directory /root/var/backups and host A,
    if /root/var/backups exist, the program will create the directory
    /root/var/backups/A.
    If /root/var/backups does not exist (or is not accessible), the
    program will log an error and exit 1.

+++ features/environment.py
@@ -56,8 +56,10 @@ def before_scenario(context, scenario):
 def after_scenario(context, scenario):
     if "notempfile" in scenario.tags:
         return
-    shutil.rmtree(context.testsource)
-    shutil.rmtree(context.testdest)
+    if os.access(context.testsource, os.F_OK):
+        shutil.rmtree(context.testsource)
+    if os.access(context.testdest, os.F_OK):
+        shutil.rmtree(context.testdest)
     os.remove(context.configfile)
 
 def after_step(context, step):
